+++ R/burrIII2.R
@@ -39,6 +39,9 @@ NULL
 #' @rdname burrIII2
 #' @export
 dburrIII2 <- function(x, lshape = 0, lscale = 1, log = FALSE) {
+  deprecate_soft("0.1.2", "xburrIII2()",
+    details = "The 'burrIII2' distribution has been deprecated for the identical 'llogis' distribution.", id = "xburrIII2"
+  )
   if (!length(x)) {
     return(numeric(0))
   }
@@ -57,6 +60,9 @@ dburrIII2 <- function(x, lshape = 0, lscale = 1, log = FALSE) {
 #' @rdname burrIII2
 #' @export
 qburrIII2 <- function(p, lshape = 0, lscale = 1, lower.tail = TRUE, log.p = FALSE) {
+  deprecate_soft("0.1.2", "xburrIII2()",
+    details = "The 'burrIII2' distribution has been deprecated for the identical 'llogis' distribution.", id = "xburrIII2"
+  )
   if (!length(q)) {
     return(numeric(0))
   }
@@ -70,6 +76,9 @@ qburrIII2 <- function(p, lshape = 0, lscale = 1, lower.tail = TRUE, log.p = FALS
 #' @rdname burrIII2
 #' @export
 pburrIII2 <- function(q, lshape = 0, lscale = 1, lower.tail = TRUE, log.p = FALSE) {
+  deprecate_soft("0.1.2", "xburrIII2()",
+    details = "The 'burrIII2' distribution has been deprecated for the identical 'llogis' distribution.", id = "xburrIII2"
+  )
   if (!length(q)) {
     return(numeric(0))
   }
@@ -82,6 +91,9 @@ pburrIII2 <- function(q, lshape = 0, lscale = 1, lower.tail = TRUE, log.p = FALS
 #' @rdname burrIII2
 #' @export
 rburrIII2 <- function(n, lshape = 0, lscale = 1) {
+  deprecate_soft("0.1.2", "xburrIII2()",
+    details = "The 'burrIII2' distribution has been deprecated for the identical 'llogis' distribution.", id = "xburrIII2"
+  )
   r <- actuar::rburr(n, shape1 = exp(lshape), shape2 = exp(lshape), scale = exp(lscale))
   1 / r
 }
@@ -89,5 +101,8 @@ rburrIII2 <- function(n, lshape = 0, lscale = 1) {
 #' @rdname burrIII2
 #' @export
 sburrIII2 <- function(x) {
+  deprecate_soft("0.1.2", "xburrIII2()",
+    details = "The 'burrIII2' distribution has been deprecated for the identical 'llogis' distribution.", id = "xburrIII2"
+  )
   list(start = list(lshape = 0, lscale = 1))
 }