+++ tests/testthat/test-aaa-deprecated.R
@@ -12,14 +12,33 @@
 #    See the License for the specific language governing permissions and
 #    limitations under the License.
 
-context("deprecated")
-
 test_that("deprecated llog", {
-  dist <- ssdtools:::ssd_fit_dist(ssdtools::boron_data, dist = "llogis")
+  dist <- ssdtools:::ssd_fit_dist(ssdtools::boron_data, dist = "llog")
 
   expect_true(is.fitdist(dist))
   expect_equal(
     coef(dist),
-    c(lscale = 0.965466010495141, lshape = -0.300741556664549)
+    c(lscale = 2.6261249, lshape = 0.7403092)
   )
+  set.seed(101)
+  pred <- predict(dist, percent = 1, ci = TRUE, nboot = 10L)
+  expect_equal(as.data.frame(pred), structure(list(
+    percent = 1, est = 0.460388430679064, se = 0.305015200817155,
+    lcl = 0.124204402582017, ucl = 1.04527103315379, dist = "llog"
+  ), row.names = c(
+    NA,
+    -1L
+  ), class = "data.frame"))
+})
+
+test_that("deprecated llog", {
+  expect_error(ssd_fit_dists(ssdtools::boron_data, dist = c("llog", "llogis")), "Distributions 'llog', 'burrIII2' and 'llogis' are identical. Please just use 'llogis'.")
+})
+
+test_that("deprecated burrIII2", {
+  expect_error(ssd_fit_dists(ssdtools::boron_data, dist = c("llog", "burrIII2")), "Distributions 'llog', 'burrIII2' and 'llogis' are identical. Please just use 'llogis'.")
+})
+
+test_that("deprecated burrIII2", {
+  expect_error(ssd_fit_dists(ssdtools::boron_data, dist = c("llogis", "burrIII2")), "Distributions 'llog', 'burrIII2' and 'llogis' are identical. Please just use 'llogis'.")
 })