+++ tests/testthat/test-data.R
@@ -12,10 +12,8 @@
 #    See the License for the specific language governing permissions and
 #    limitations under the License.
 
-context("data")
-
 test_that("data", {
-  expect_identical(checkr::check_data(
+  expect_null(chk::check_data(
     ccme_data,
     values = list(
       Chemical = "",
@@ -24,11 +22,11 @@ test_that("data", {
       Conc = c(0, Inf),
       Group = factor(c("Amphibian", "Fish", "Invertebrate", "Plant"))
     ),
-    nrow = c(1, Inf)
-  ), ccme_data)
+    nrow = c(1L, 2147483647L)
+  ))
   expect_is(ccme_data, "tbl")
 
-  expect_identical(checkr::check_data(
+  expect_null(chk::check_data(
     boron_data,
     values = list(
       Chemical = c("Boron", "Boron"),
@@ -38,10 +36,10 @@ test_that("data", {
       Group = factor(c("Amphibian", "Fish", "Invertebrate", "Plant"))
     ),
     nrow = 28
-  ), boron_data)
+  ))
   expect_is(boron_data, "tbl")
 
-  expect_identical(checkr::check_data(
+  expect_null(chk::check_data(
     boron_pred,
     values = list(
       percent = 1L,
@@ -51,9 +49,21 @@ test_that("data", {
       ucl = 1
     ),
     nrow = 99L
-  ), boron_pred)
+  ))
   expect_is(boron_pred, "tbl")
 
   expect_is(boron_lnorm, "fitdist")
   expect_is(boron_dists, "fitdists")
+
+  expect_null(chk::check_data(
+    test_data,
+    values = list(
+      Chemical = "",
+      Conc = c(0.04, 76500)
+    ),
+    exclusive = TRUE,
+    order = TRUE,
+    nrow = 141L
+  ))
+  expect_is(test_data, "tbl")
 })