+++ tests/testthat/test-hc.R
@@ -12,8 +12,6 @@
 #    See the License for the specific language governing permissions and
 #    limitations under the License.
 
-context("hc")
-
 test_that("ssd_hc list", {
   expect_error(ssd_hc(list()), "^`x` must be named[.]$", class = "chk_error")
 
@@ -35,7 +33,10 @@ test_that("ssd_hc list", {
     structure(list(
       percent = 5, est = 0.193040816698737, se = NA_real_,
       lcl = NA_real_, ucl = NA_real_, dist = "lnorm"
-    ), row.names = "lnorm", class = "data.frame")
+    ), row.names = c(
+      NA,
+      -1L
+    ), class = "data.frame")
   )
   expect_equal(
     as.data.frame(ssd_hc(list("lnorm" = NULL), percent = c(1, 99))),
@@ -45,7 +46,7 @@ test_that("ssd_hc list", {
     ), se = c(NA_real_, NA_real_), lcl = c(
       NA_real_,
       NA_real_
-    ), ucl = c(NA_real_, NA_real_), dist = c("lnorm", "lnorm")), row.names = c("lnorm.1", "lnorm.2"), class = "data.frame")
+    ), ucl = c(NA_real_, NA_real_), dist = c("lnorm", "lnorm")), row.names = c(NA, -2L), class = "data.frame")
   )
 
   expect_equal(
@@ -53,7 +54,10 @@ test_that("ssd_hc list", {
     structure(list(
       percent = 5, est = 0.193040816698737, se = NA_real_,
       lcl = NA_real_, ucl = NA_real_, dist = "lnorm"
-    ), row.names = "lnorm", class = "data.frame")
+    ), row.names = c(
+      NA,
+      -1L
+    ), class = "data.frame")
   )
 
   expect_equal(
@@ -61,7 +65,10 @@ test_that("ssd_hc list", {
     structure(list(
       percent = 5, est = 0.275351379333677, se = NA_real_,
       lcl = NA_real_, ucl = NA_real_, dist = "lnorm"
-    ), row.names = "lnorm", class = "data.frame")
+    ), row.names = c(
+      NA,
+      -1L
+    ), class = "data.frame")
   )
 
   expect_equal(
@@ -72,7 +79,7 @@ test_that("ssd_hc list", {
     ), se = c(NA_real_, NA_real_), lcl = c(
       NA_real_,
       NA_real_
-    ), ucl = c(NA_real_, NA_real_), dist = c("lnorm", "llogis")), row.names = c("lnorm", "llogis"), class = "data.frame")
+    ), ucl = c(NA_real_, NA_real_), dist = c("lnorm", "llogis")), row.names = c(NA, -2L), class = "data.frame")
   )
 
   expect_equal(
@@ -89,10 +96,7 @@ test_that("ssd_hc list", {
     ), ucl = c(NA_real_, NA_real_, NA_real_, NA_real_), dist = c(
       "lnorm",
       "lnorm", "llogis", "llogis"
-    )), row.names = c(
-      "lnorm.1", "lnorm.2",
-      "llogis.1", "llogis.2"
-    ), class = "data.frame")
+    )), row.names = c(NA, -4L), class = "data.frame")
   )
 })
 
@@ -127,7 +131,7 @@ test_that("ssd_hc fitdists", {
   expect_equal(
     as.data.frame(ssd_hc(boron_dists)),
     structure(list(
-      percent = 5, est = 1.30474651622516, se = NA_real_,
+      percent = 5, est = 1.30686521357099, se = NA_real_,
       lcl = NA_real_, ucl = NA_real_, dist = "average"
     ), class = "data.frame", row.names = c(
       NA,
@@ -140,7 +144,7 @@ test_that("ssd_hc fitdists not average", {
   expect_equal(
     as.data.frame(ssd_hc(boron_dists, average = FALSE)),
     structure(list(percent = c(5, 5, 5), est = c(
-      1.58920212463066,
+      1.56256632555312,
       1.07373870642628, 1.68117483775796
     ), se = c(
       NA_real_, NA_real_,
@@ -148,7 +152,7 @@ test_that("ssd_hc fitdists not average", {
     ), lcl = c(NA_real_, NA_real_, NA_real_), ucl = c(
       NA_real_,
       NA_real_, NA_real_
-    ), dist = c("burrIII2", "gamma", "lnorm")), row.names = c(
+    ), dist = c("llogis", "gamma", "lnorm")), row.names = c(
       NA,
       -3L
     ), class = "data.frame")
@@ -159,7 +163,7 @@ test_that("ssd_hc fitdistscens", {
   expect_equal(
     as.data.frame(ssd_hc(fluazinam_dists)),
     structure(list(
-      percent = 5, est = 1.35230977078523, se = NA_real_,
+      percent = 5, est = 1.42068088873316, se = NA_real_,
       lcl = NA_real_, ucl = NA_real_, dist = "average"
     ), class = "data.frame", row.names = c(
       NA,
@@ -172,7 +176,7 @@ test_that("ssd_hc fitdistscens not average", {
   expect_equal(
     as.data.frame(ssd_hc(fluazinam_dists, average = FALSE)),
     structure(list(percent = c(5, 5, 5), est = c(
-      1.11030265567844,
+      1.30938169835089,
       0.309067069393034, 1.74352219048516
     ), se = c(
       NA_real_, NA_real_,
@@ -180,7 +184,7 @@ test_that("ssd_hc fitdistscens not average", {
     ), lcl = c(NA_real_, NA_real_, NA_real_), ucl = c(
       NA_real_,
       NA_real_, NA_real_
-    ), dist = c("burrIII2", "gamma", "lnorm")), row.names = c(
+    ), dist = c("llogis", "gamma", "lnorm")), row.names = c(
       NA,
       -3L
     ), class = "data.frame")