+++ tests/testthat/test-predict.R
@@ -12,17 +12,18 @@
 #    See the License for the specific language governing permissions and
 #    limitations under the License.
 
-context("predict")
-
 test_that("predict.fitdist", {
+  rlang::scoped_options(lifecycle_verbosity = "quiet")
+  
   boron_lnorm <- ssdtools:::ssd_fit_dist(ssdtools::boron_data[1:6, ])
   pred <- predict(boron_lnorm, nboot = 10L)
+  
   expect_is(pred, "tbl")
   expect_identical(colnames(pred), c("percent", "est", "se", "lcl", "ucl", "dist"))
   expect_equal(pred$percent, 1:99)
   pred2 <- predict(boron_lnorm, ci = TRUE, nboot = 10)
   expect_identical(pred$est[1], pred2$est[1])
-
+  
   boron_data$Conc <- boron_data$Conc / 1000
   boron_lnorm3 <- ssdtools:::ssd_fit_dist(boron_data[1:6, ])
   pred3 <- predict(boron_lnorm3, nboot = 10)
@@ -30,8 +31,9 @@ test_that("predict.fitdist", {
 })
 
 test_that("predict.fitdist parallel", {
+  rlang::scoped_options(lifecycle_verbosity = "quiet")
   boron_lnorm <- ssdtools:::ssd_fit_dist(ssdtools::boron_data)
-
+  
   pred <- predict(boron_lnorm, nboot = 10L, parallel = "multicore", ncpus = 2)
   expect_is(pred, "tbl")
   expect_identical(colnames(pred), c("percent", "est", "se", "lcl", "ucl", "dist"))
@@ -39,12 +41,14 @@ test_that("predict.fitdist parallel", {
 })
 
 test_that("predict.fitdists", {
+  rlang::scoped_options(lifecycle_verbosity = "quiet")
   dists <- ssd_fit_dists(boron_data[1:6, ], dists = c("gamma", "gompertz"))
   pred <- predict(dists, nboot = 10L)
+  
   expect_is(pred, "tbl")
   expect_identical(colnames(pred), c("percent", "est", "se", "lcl", "ucl", "dist"))
   expect_equal(pred$percent, 1:99)
-
+  
   pred <- predict(dists, average = FALSE)
   expect_is(pred, "tbl")
   expect_identical(colnames(pred), c("percent", "est", "se", "lcl", "ucl", "dist"))
@@ -53,8 +57,9 @@ test_that("predict.fitdists", {
 })
 
 test_that("predict.fitdists parallel", {
+  rlang::scoped_options(lifecycle_verbosity = "quiet")
   boron_lnorm <- ssd_fit_dists(ssdtools::boron_data, dists = c("gamma", "gompertz"))
-
+  
   pred <- predict(boron_lnorm, nboot = 10L, parallel = "multicore", ncpus = 2)
   expect_is(pred, "tbl")
   expect_identical(colnames(pred), c("percent", "est", "se", "lcl", "ucl", "dist"))
@@ -62,33 +67,37 @@ test_that("predict.fitdists parallel", {
 })
 
 test_that("predict.fitdistscens", {
+  rlang::scoped_options(lifecycle_verbosity = "quiet")
+  pred <- predict(ssdtools::fluazinam_dists, percent = c(1, 99))
   expect_equal(
-    as.data.frame(predict(ssdtools::fluazinam_dists, percent = c(1, 99))),
+    as.data.frame(pred),
     structure(list(percent = c(1, 99), est = c(
-      0.165191855589344,
-      74931.0016372917
+      0.17476496294754,
+      126041.188199444
     ), se = c(NA_real_, NA_real_), lcl = c(
       NA_real_,
       NA_real_
-    ), ucl = c(NA_real_, NA_real_), dist = c("average", "average")), class = "data.frame", row.names = c(NA, -2L))
+    ), ucl = c(NA_real_, NA_real_), dist = c("average", "average")), row.names = c(NA, -2L), class = "data.frame")
   )
 })
 
 test_that("predict.fitdistscens cis", {
   set.seed(77)
   pred <- predict(ssdtools::fluazinam_dists,
-    percent = c(1, 99),
-    ci = TRUE, average = FALSE, nboot = 10
+                  percent = c(1, 99),
+                  ci = TRUE, average = FALSE, nboot = 10
   )
   expect_identical(colnames(pred), c("percent", "est", "se", "lcl", "ucl", "dist"))
   expect_identical(pred$percent, c(1, 99, 1, 99, 1, 99))
-  expect_equal(pred$est, c(
-    0.0556070303830483,
-    93128.5004982232, 0.00297074136543809, 6884.80066910368, 0.279206726612854,
-    75330.7588691179
-  ))
+  expect_equal(
+    pred$est,
+    c(
+      0.0947064911703993, 212423.73904855, 0.00297074136543809, 6884.80066910368,
+      0.279206726612854, 75330.7588691179
+    )
+  )
   expect_identical(pred$dist, c(
-    "burrIII2", "burrIII2", "gamma",
+    "llogis", "llogis", "gamma",
     "gamma", "lnorm", "lnorm"
   ))
 })